requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
alembic
redis
qrcode[pil]
pydantic-settings
python-multipart
asyncpg
